refactor: detail이랑 list 관련 폴더 분리
