+++ src/features/contentList/api/queries.ts
@@ -6,8 +6,6 @@ import {
 import {
   GetCollegeResData,
   GetDepartmentResData,
-  GetEventDetailResData,
-  GetNoticeDetailResData,
   GetNoticeResData,
 } from './types';
 import { AxiosError } from 'axios';
@@ -80,18 +78,6 @@ export const useGetInfiniteCampusNotice = (campusNoticeType: string) => {
   });
 };
 
-//학교 공지 detail
-export const useGetCampusNoticeDetail = (
-  campusNoticeId: number,
-  options?: UseQueryOptions<GetNoticeDetailResData, AxiosError>,
-) => {
-  return useQuery({
-    queryKey: ['campusNoticeDetail', campusNoticeId],
-    queryFn: () => NOTICE_API.campusNoticeDetail(campusNoticeId),
-    ...options,
-  });
-};
-
 //자치단체 공지
 export const useGetNotice = (affiliation: string) => {
   return useInfiniteQuery({
@@ -113,21 +99,6 @@ export const useGetNotice = (affiliation: string) => {
   });
 };
 
-//자치단체 공지 detail
-export const useGetNoticeDetail = (
-  noticeId: number,
-  options?: UseQueryOptions<GetNoticeDetailResData, AxiosError>,
-) => {
-  return useQuery({
-    queryKey: ['campusNoticeDetail'],
-    queryFn: async () => {
-      const data = await NOTICE_API.noticeDetail(noticeId);
-      return data;
-    },
-    ...options,
-  });
-};
-
 export const useGetEvent = (affiliation: string) => {
   return useInfiniteQuery({
     queryKey: ['event', affiliation],
@@ -145,20 +116,6 @@ export const useGetEvent = (affiliation: string) => {
       if (lastPage.content.length < 6) return undefined;
       return lastPage.number + 1;
     },
-  });
-};
-
-export const useGetEventDetail = (
-  eventId: number,
-  options?: UseQueryOptions<GetEventDetailResData, AxiosError>,
-) => {
-  return useQuery({
-    queryKey: ['eventDetail', eventId],
-    queryFn: async () => {
-      const data = await EVENT_API.eventDetail(eventId);
-      return data;
-    },
-    ...options,
   });
 };
